The speed of the Instagram bot has been increased and optimized a lot

+++ handlers/groups/group_handler.py
@@ -1,10 +1,12 @@
+import re
+import time
 from aiogram.dispatcher import FSMContext
 from aiogram.types import InputMediaPhoto, InputMediaVideo
-
+import keyboards
 from data import *
 from aiogram import types
+from handlers import instagram_api
 from keyboards import *
-from download import *
 from utlis.models import *
 
 
@@ -61,54 +63,60 @@ async def process_language_selection(call: types.CallbackQuery, state: FSMContex
         logger.exception("Error while processing language selection: %s", e)
 
 
-@dp.message_handler(regexp=r'https?:\/\/(www\.)?instagram\.com\/(reel|p|tv)\/([-_a-zA-Z0-9]{11})')
+@dp.message_handler(regexp=r'https?:\/\/(www\.)?instagram\.com\/(reel|p)\/([-_a-zA-Z0-9]{11})')
 async def send_instagram_media(message: types.Message):
-    global waiting_msg
+    global waiting_msg, delete_msg
     link = message.text
     language = await Group.get_language(message.chat.id)
+    await message.delete()
+    match = re.search(r'https://www.instagram.com/(?:p|reel)/([-_a-zA-Z0-9]{11})', link)
+    vid = match.group(1) if match else None
     try:
-        await message.delete()
+        cached_data = instagram_api.cache.get(vid, {})
+        if cached_data.get('timestamp', 0) >= time.time() - 2629746:
+            media = cached_data.get('result')
+            return await bot.send_media_group(chat_id=message.chat.id, media=media)
         waiting_msg = await bot.send_message(chat_id=message.chat.id,
-                                             text=f"<b>📥 {keyboard_waiting[language]}</b>", protect_content=True)
-        urls = await download_media(link=link)
+                                             text=f"<b>📥 {keyboards.keyboard_waiting[language]}</b>",
+                                             protect_content=True)
+        urls = await instagram_api.instagram_downloader(vid=vid)
         if urls is None or not urls:
             await waiting_msg.delete()
-            return await bot.send_message(message.chat.id, text=down_err[language].format(link),
-                                          disable_web_page_preview=True)
+            return await bot.send_message(message.chat.id, text=keyboards.down_err[language].format(link),
+                                          disable_web_page_preview=True, protect_content=True)
         media = [InputMediaPhoto(url) if 'jpg' in url else InputMediaVideo(url) for url in urls]
-        media[
-            -1].caption = f"<a href='tg://user?id={message.from_user.id}'><b>↬  υ𝐬𝐞𝐫</b></a>\n\n<b>📥 {main_caption}{keyboard_saver[language]}</b>"
+        media[-1].caption = f"<b>📥 {main_caption}{keyboards.keyboard_saver[language]}</b>"
         await bot.send_media_group(chat_id=message.chat.id, media=media)
         await waiting_msg.delete()
+        instagram_api.cache[vid] = {'result': media, 'timestamp': time.time()}
     except Exception as e:
         await waiting_msg.delete()
         logger.exception("Error while sending Instagram photo: %s", e)
-        return await bot.send_message(message.chat.id, text=down_err[language].format(link),
+        return await bot.send_message(message.chat.id, text=keyboards.down_err[language].format(link),
                                       disable_web_page_preview=True, protect_content=True)
 
 
-@dp.message_handler(regexp=r'https?:\/\/(www\.)?instagram\.com\/(stories)', chat_type=types.ChatType.PRIVATE)
+@dp.message_handler(regexp=r'https?:\/\/(www\.)?instagram\.com\/(stories)')
 async def send_instagram_media(message: types.Message):
     link = message.text
     await message.delete()
     language = await Group.get_language(message.chat.id)
     waiting_msg = await bot.send_message(chat_id=message.chat.id,
-                                         text=f"<b>📥 {keyboard_waiting[language]}</b>", protect_content=True)
+                                         text=f"<b>📥 {keyboards.keyboard_waiting[language]}</b>", protect_content=True)
     try:
-        urls = await download_media(link=link)
+        urls = await instagram_api.instagram_stories(link=link)
         if urls is None or not urls:
             await waiting_msg.delete()
-            return await bot.send_message(message.chat.id, text=down_err[language].format(link),
-                                          disable_web_page_preview=True)
+            return await bot.send_message(message.chat.id, text=keyboards.down_err[language].format(link),
+                                          disable_web_page_preview=True, protect_content=True)
         media_groups = [urls[i:i + 10] for i in range(0, len(urls), 10)]
         for group in media_groups:
             media = [InputMediaPhoto(url) if 'jpg' in url else InputMediaVideo(url) for url in group]
-            media[
-                -1].caption = f"<a href='tg://user?id={message.from_user.id}'><b>↬  υ𝐬𝐞𝐫</b></a>\n\n<b>📥 {main_caption}{keyboard_saver[language]}</b>"
+            media[-1].caption = f"📥 <b>{main_caption}{keyboards.keyboard_saver[language]}</b>"
             await bot.send_media_group(chat_id=message.chat.id, media=media)
         await waiting_msg.delete()
     except Exception as e:
-        logger.exception("Error while sending Instagram photo: %s", e)
+        logger.exception("Error while sending Instagram photo: %s", )
         await waiting_msg.delete()
-        return await bot.send_message(message.chat.id, text=down_err[language].format(link),
+        return await bot.send_message(message.chat.id, text=keyboards.down_err[language].format(link),
                                       disable_web_page_preview=True, protect_content=True)
